Upgrade gulp-sourcemaps
Last week, we pinned gulp-sourcemaps to version 1.6.0 to avoid a bug in
a new version.[1] That issue has been fixed, so we can now go back to
using the flexible ^ version selector.

gulp-sourcemaps 2 is now available and it works well with our existing
code, so we can go and update straight to that.

[1] https://github.com/floridoo/gulp-sourcemaps/issues/243