Show example data, allow cohort configuration
The front-end is updated to show example data from Experiment 1. The
experiment is not yet configurable; that's coming later.

The user can choose which cohorts should be shown on the charts. The set
of possible cohorts is fetched dynamically from the back-end, since each
experiment has a different set of possible cohorts.

Under the hood, cohorts are what Distribution Viewer called populations.
Rather than use the term "cohorts" internally, I use the more general term
"subgroups". So when you see the term "subgroups", think "cohorts a.k.a.
populations".