Fix #55: Don't change colors after disabling pops
Before this change, populations would be colored based only on their
source order. This worked for the most part, but there was one important
problem with this approach.

For example, imagine two populations were enabled: A and B. A would be
blue and B would be purple. Upon disabling the A population, the B
population would be first in source order and would thus be blue. In
other words, B would take on A's color. That was confusing.

This commit fixes that.