Fix #131: Draw lines for all non-empty populations
Before this change, the function that returned population line
components would bail out whenever an empty population was encountered.

The bug stems from an oversight I made while copying a for loop from
another part of the code. Now that the for loop is inside a function, it
shouldn't return null when encountering an empty population but instead
simply skip over that population.