Issue #15: Add package-lock.json
This is something that came up while working on #15. I could split it
into another commit but then I would need to wait for this one to land.

This file is generated automatically in NPM 5 and the team recommends
committing it. NPM has actually had lock files forever with the
"shrinkwrap" feature, but this is surfaced better and officially
recommended.