Rebuild the front-end during deployment
When an application is promoted from staging to production, Heroku does
not rebuild that application in the production environment. This can be
a problem if any of the build processes involve environment variables
that differ between staging and production.

Quoting the Heroku documentation:

    When promoting slugs between apps, Heroku will copy the slug
    verbatim between apps, without changes. In particular, the slug is
    not rebuilt with the environment of the target app. If your builds
    tend to bake in environment from the build-app context, then your
    slugs will not be portable between apps, and will not work correctly
    when promoted from the app where they were built. This may be the
    case, for example, if you’re using Ruby on Rails and the asset
    pipeline to build assets hosted at a CDN defined by a URL in your
    build environment. This is because URLs specific to the build-app
    will be baked into css and js-files, and those will not work or not
    work correctly when promoted.

    - https://devcenter.heroku.com/articles/pipelines

In Experiments Viewer, TRACKING_ID is set on production but not staging.
The value of TRACKING_ID is read at build time, not runtime, so we need
to rebuild the application when it's promoted to production.

In this commit, we use the Release Phase feature of Heroku to do that:
https://devcenter.heroku.com/articles/release-phase