Remove release line from Procfile
As it turns out, this doesn't solve the problem we were having.

When promoting an application, the release task is run in the
environment of the source dyno, not the target dyno. So it doesn't solve
the problem of TRACKING_ID only being available on production.

To get the front-end to build correctly for production, we need to set
TRACKING_ID to the same value on staging.