Revert striptags to 2.2.1
The latest version of striptags is causing the build task to fail
because it uses ES6, which is not supported by the version of Uglify
that Webpack is using.

As soon as [1] is published to a main release and Webpack updates to
use that release, we should be able to upgrade striptags to the latest
version.

[1] https://github.com/webpack-contrib/uglifyjs-webpack-plugin/pull/63