Proactively upgrade all NPM packages
This is just easier than merging the Renovate PRs one-by-one. This
change disables React deprecation warnings in eslint, but that's only
because this project is going to redirect to Test Tube soon. I'll open
an issue to address these deprecation warnings in the unlikely case that
we don't redirect to Test Tube.